requirements.txt
# Core TTS Libraries
chatterbox-tts>=0.1.0
torch>=2.5.1
torchaudio>=2.5.1

# Audio Processing
scipy>=1.10.0
librosa>=0.10.0

# Utilities
numpy>=1.24.0
requests>=2.28.0

# Optional: For Bark TTS
# git+https://github.com/suno-ai/bark.git

# Optional: For XTTS-v2
# TTS>=0.22.0 